Fixed potential issues related to comparing ints to longs

+++ Px.Utils/Operations/MatrixFunctionExtensions.cs
@@ -32,7 +32,7 @@ namespace Px.Utils.Operations
             IMatrixMap resultOnlyMap = newMeta.CollapseDimension(sourceMap.Code, newValue.Code);
             long arraySize = input.Metadata.GetSizeLong() + resultOnlyMap.GetSizeLong();
             if (arraySize > int.MaxValue) throw new InvalidOperationException("Resulting matrix size exceeds maximum array size.");
-            TData[] outData = new TData[arraySize];
+            TData[] outData = new TData[(int)arraySize];
 
             // Initialize the output matrix with the identity value
             for (int i = 0; i < outData.Length; i++) outData[i] = functionIdentity;
